some report are working

+++ resources/js/components/allowance_type.vue
@@ -1,11 +1,14 @@
 <template>
 	<div>
-		<h1>Add allowance types</h1>
+		
+		
 
 		<form action="#" @submit.prevent="edit ? updateAllowanceType(allowance_type.id): createAllowanceType()">
 
-			<div class="form-group">				
-				<input v-model="allowance_type.name" type="text" name="name" class="form-control" placeholder="Name">
+			<div class="form-group">
+								
+				 <input v-model="allowance_type.name" type="text" name="name" class="form-control" placeholder="Name">
+		
 			</div>
 
 			<div class="form-group">				
@@ -19,40 +22,80 @@
 
 
 		</form>
-		<h1>Allowance types</h1>
-		<ul class="list-group">
-			<li class="list-group-item" v-for="allowance_type in list">
-				{{ allowance_type.id }} | {{ allowance_type.name }} | {{ allowance_type.description }} |
-				<button @click="showAllowanceType(allowance_type.id)"  class="btn btn-default btn-xs">edit</button>|
-				<button @click="deleteAllowanceType(allowance_type.id)"  class="btn btn-danger btn-xs">delete</button>
 
-				
-			</li>
-				
-			
-		</ul>
+		
+
+		
+		<div class="table-responsive">
+
+              <table class="table  table-hover table-striped table-bordered table-sm">   
+
+                  <thead>
+	<tr>
+		<th>id</th>
+		<th>Name</th>
+		<th>Description</th>
+	
+		<th>Edit</th>
+		<th>Actions</th>
+	</tr>
+                  </thead>
+                  <tbody>
+					        <tr class="" v-if="list.length === 0">
+        <td class="lead text-center" :colspan="4">No data found.</td>
+      </tr>
+
+     
+                    <tr  v-for="allowance_type in list" :key="allowance_type.id">
+                      <td>{{allowance_type.id }}</td>
+
+                    <td>{{ allowance_type.name }}</td>
+
+					<td>{{ allowance_type.description }}</td>
+
+                    <td><button @click="showAllowanceType(allowance_type.id)"  class="btn btn-default btn-sm">edit</button></td>
+
+					
+
+                    <td><button @click="deleteAllowanceType(allowance_type.id)"  class="btn btn-danger btn-sm">delete</button>
+                    </td>
+
+                    </tr>
+     
+
+        </tbody>
+      </table>
+  </div>
+    
+
 	</div>
 </template>
 
 <script>
 
 	export default{
-		data: function(){
+		data: function(){			
 			return {
+				renderComponent: true,
 				edit:false,
 				list:[],
 				allowance_type: {
 					id:'',
 					name:'',
 					description:''
-				}				
+				},			
+				
 			}
 		},
+
+		created() {
+			this.fetchAllowanceTypeList();		
+		},
+
 
 		mounted: function(){
 			console.log('allowance types component loaded ...');
 			this.fetchAllowanceTypeList();
-
 		},
 		
 
@@ -121,6 +164,9 @@
 			},
 
 			deleteAllowanceType: function(id){
+
+             if(confirm("Do you really want to delete "+this.id+ "? ")){
+
 				console.log('deleting allowance_type....');	
 				let self = this;			
 				axios.delete('/api/allowance_types/'+id).
@@ -129,10 +175,16 @@
 					})
 					.catch(function(error){
 						console.log(error);
-					});			
+					});
+			 }	
+			 	
 
-			}
-		}
+			},
+
+	
+        }
+
+		
 	}
 	
 </script>
